Try to open /dev/mem before calling rpitx
mailbox.c tries to open /dev/mem and calls exit if it doesn't have
permission, which will forcefully exit the Python runtime. Try to open
it and throwing an exception if it fails before umping into C to give
users a chance to recover.

Also, because avconv is writing to a pipe, having the read end of the
pipe unexpectedly close breaks the terminal. I kept having to run
`reset` to get it back.